packages/payai/requirements.txt
# Kamiyo - Exploit Intelligence Aggregator
# Requirements

# Core dependencies
feedparser==6.0.11
python-dateutil==2.9.0
requests==2.32.3
PyYAML==6.0.2

# API dependencies
fastapi==0.115.0
uvicorn==0.32.1
pydantic==2.10.3
email-validator>=2.0.0  # Required by pydantic for email validation
httpx  # Version auto-resolved for compatibility

# Database - PostgreSQL (Production)
psycopg2-binary==2.9.9
sqlalchemy==2.0.23  # Optional ORM

# Database - SQLite (Development)
# Built-in to Python

# Caching
redis==5.0.1
hiredis==2.2.3

# Payment Processing
stripe>=7.0.0  # Day 6: Stripe API integration for subscription billing

# Authentication
pyjwt==2.8.0
bcrypt==4.1.2
passlib==1.7.4

# Email
sendgrid==6.11.0
jinja2==3.1.2

# Monitoring & Logging
prometheus-client==0.19.0
sentry-sdk==1.39.2
python-json-logger==2.0.7
structlog==24.1.0  # Structured logging for social media monitoring
psutil==5.9.8  # System metrics for health checks

# Rate Limiting
slowapi==0.1.9  # Rate limiting for FastAPI

# Retry Logic
tenacity==8.2.3  # Retry logic for API calls and external services

# Scheduling
schedule==1.2.2
celery==5.3.4  # For background tasks

# Web3 & Blockchain
web3==6.11.3  # On-chain monitoring (Ethereum, Base, Arbitrum)
solana>=0.34.0  # Solana blockchain interaction for x402 payments (updated for httpx compatibility)

# Twitter/Social
tweepy==4.14.0
beautifulsoup4==4.12.2

# Telegram Integration (Day 25)
python-telegram-bot==20.7
telethon>=1.36.0  # Telegram monitoring for Phase 1 (production-ready)

# Discord Integration (Day 26)
discord.py==2.3.2  # Discord bot library with slash command support

# Social Media Posting (Day 31)
praw==7.7.1  # Reddit API wrapper
# tweepy already listed above (line 52)
websockets>=10.0,<12.0  # WebSocket client, compatible with solana and web3

# AI Enhancement for Deep Dive Analysis
anthropic>=0.34.0  # Claude AI for enhanced thread generation

# Testing
pytest==8.3.4
pytest-cov==4.1.0
pytest-asyncio==0.23.2

# Development
black==24.8.0
flake8==6.1.0
mypy==1.7.1

# Security
python-dotenv==1.0.0
cryptography==41.0.7
fastapi-csrf-protect==0.3.4  # Python 3.10+ required (now using Python 3.11.14)

# Performance & Optimization (Days 14-15)
brotli==1.1.0  # Brotli compression (better than gzip)
pybloom-live==4.0.0  # Bloom filter for deduplication
aiodns==3.1.1  # Async DNS resolution for aggregators
charset-normalizer==3.3.2  # Fast character encoding detection (replaces cchardet, Python 3.11+ compatible)
click==8.1.7

# Phase 1 Dependencies - On-chain & Forta Monitoring
gql>=3.5.0  # GraphQL client for Forta Network API
aiohttp>=3.9.0  # Async HTTP client for Forta API calls
python-Levenshtein>=0.25.0  # Fuzzy string matching for deduplication
